Move type to own file.
